ref(bug reports): turn read, unread, and delete into hamburger dropdown
