+++ static/app/components/feedback/feedbackItem/feedbackItem.tsx
@@ -1,9 +1,16 @@
-import {Fragment} from 'react';
+import {Fragment, useCallback, useMemo} from 'react';
 import styled from '@emotion/styled';
 
+import {
+  addErrorMessage,
+  addLoadingMessage,
+  addSuccessMessage,
+} from 'sentry/actionCreators/indicator';
 import {CopyToClipboardButton} from 'sentry/components/copyToClipboardButton';
+import {DropdownMenu} from 'sentry/components/dropdownMenu';
 import ErrorBoundary from 'sentry/components/errorBoundary';
-import DeleteButton from 'sentry/components/feedback/feedbackItem/deleteButton';
+import {useInfiniteFeedbackListData} from 'sentry/components/feedback/feedbackDataContext';
+import openDeleteModal from 'sentry/components/feedback/feedbackItem/deleteButton';
 import Section from 'sentry/components/feedback/feedbackItem/feedbackItemSection';
 import FeedbackItemUsername from 'sentry/components/feedback/feedbackItem/feedbackItemUsername';
 import FeedbackViewers from 'sentry/components/feedback/feedbackItem/feedbackViewers';
@@ -14,11 +21,13 @@ import ObjectInspector from 'sentry/components/objectInspector';
 import PanelItem from 'sentry/components/panels/panelItem';
 import {Flex} from 'sentry/components/profiling/flex';
 import TextCopyInput from 'sentry/components/textCopyInput';
-import {IconJson, IconLink} from 'sentry/icons';
+import {IconEllipsis, IconJson, IconLink} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {HydratedFeedbackItem} from 'sentry/utils/feedback/item/types';
+import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
+import useUrlParams from 'sentry/utils/useUrlParams';
 
 interface Props {
   feedbackItem: HydratedFeedbackItem;
@@ -26,6 +35,31 @@ interface Props {
 
 export default function FeedbackItem({feedbackItem}: Props) {
   const organization = useOrganization();
+  const api = useApi();
+  const {getParamValue: getFeedbackSlug, setParamValue: setFeedbackSlug} =
+    useUrlParams('feedbackSlug');
+  const {setFeedback} = useInfiniteFeedbackListData();
+  const feedbackId = feedbackItem.feedback_id;
+
+  const url = useMemo(() => {
+    const feedbackSlug = getFeedbackSlug();
+    const projectSlug = feedbackSlug?.split(':')[0];
+    return `/projects/${organization.slug}/${projectSlug}/feedback/${feedbackId}/`;
+  }, [feedbackId, getFeedbackSlug, organization]);
+
+  const handleDelete = useCallback(async () => {
+    addLoadingMessage(t('Deleting feedback...'));
+    try {
+      await api.requestPromise(url, {method: 'DELETE'});
+      addSuccessMessage(t('Deleted feedback'));
+      setFeedbackSlug('');
+      setFeedback(feedbackId, undefined);
+    } catch {
+      addErrorMessage(t('An error occurred while deleting the feedback.'));
+    }
+  }, [api, feedbackId, setFeedback, setFeedbackSlug, url]);
+
+  const onDelete = () => openDeleteModal({onDelete: handleDelete});
 
   return (
     <Fragment>
@@ -50,7 +84,32 @@ export default function FeedbackItem({feedbackItem}: Props) {
               <ResolveButton feedbackItem={feedbackItem} />
             </ErrorBoundary>
             <ErrorBoundary mini>
-              <DeleteButton feedbackItem={feedbackItem} />
+              <DropdownMenu
+                position="bottom-end"
+                triggerProps={{
+                  'aria-label': t('Feedback Actions Menu'),
+                  icon: <IconEllipsis size="xs" />,
+                  showChevron: false,
+                  size: 'xs',
+                }}
+                items={[
+                  {
+                    key: 'mark read',
+                    label: t('Mark as read'),
+                    onAction: () => {},
+                  },
+                  {
+                    key: 'mark unread',
+                    label: t('Mark as unread'),
+                    onAction: () => {},
+                  },
+                  {
+                    key: 'delete',
+                    label: t('Delete'),
+                    onAction: onDelete,
+                  },
+                ]}
+              />
             </ErrorBoundary>
           </Flex>
         </Flex>
